Add images and new recipe. Finish first recipe.
Finished writing the pretzel-bread.html page by adding an image, ingredients
and instructions. Also created a new recipe, lime-in-coconut.html,
containing an image and a brief, sarcastic heavy, recipe. Links back to
index.html were added to the bottom of both recipe pages. Index.html was
also updated with a more detailed description of the pretzel bread recipe,
as well as a description of the lime-in-coconut recipe and the homepage
itself.